refactor: extract scheduleSchema to src/schedule.ts
Move scheduleSchema to a dedicated module so tests can import it
without triggering RedashClient initialization, which requires
REDASH_URL and REDASH_API_KEY environment variables.
